requirements.txt
fastapi
resend
python-dotenv

uvicorn
